+++ src/components/ui/index.ts
@@ -4,3 +4,4 @@ export { default as IconButton } from "./IconButton";
 export { default as Input } from "./Input";
 export { default as Modal } from "./Modal";
 export { default as Textarea } from "./Textarea";
+export { default as Tooltip } from "./Tooltip";